combined array algorithms in one section
