+++ algo/tests/array_tester.py
@@ -26,6 +26,60 @@ def arrayAlgorthims() -> None:
 
     removeDuplicatesTest([1,1,2])
     removeDuplicatesTest([0,0,1,1,1,2,2,3,3,4])
+    thinline()
+
+    pivotIndexTest([1,7,3,6,5,6])
+    pivotIndexTest([1,2,3])
+    pivotIndexTest([2,1,-1])
+    thinline()
+
+    dominantIndexTest([3,6,1,0])
+    dominantIndexTest([1,2,3,4])
+    thinline()
+
+    plusOneTest([1,2,3])
+    plusOneTest([4,3,2,1])
+    plusOneTest([9])
+    thinline()
+
+    # findDiagonalOrderTest([[1,2,3],[4,5,6],[7,8,9]])
+    # findDiagonalOrderTest[[1,2],[3,4]]
+    # thinline()
+
+    spiralOrderTest([[1,2,3],[4,5,6],[7,8,9]])
+    spiralOrderTest([[1,2,3,4],[5,6,7,8],[9,10,11,12]])
+    spiralOrderTest([[3],[2]])
+    thinline()
+
+    generatePascalTest(5)
+    generatePascalTest(1)
+    thinline()
+
+    binarySumTest("11", "1")
+    binarySumTest("1010", "1011")
+    thinline()
+
+    findStrTest("hello", "ll")
+    findStrTest("aaaaa", "bba")
+    findStrTest("a", "a")
+    findStrTest("abc", "c")
+    thinline()
+
+    longestCommonPrefixTest(["flower","flow","flight"])
+    longestCommonPrefixTest(["dog","racecar","car"])
+    thinline()
+
+    reverseStringTest(['h','e','l','l','o'])
+    reverseStringTest(['H','a','n','n','a','h'])
+    thinline()
+
+    arrayPairSumTest([1,4,3,2])
+    arrayPairSumTest([6,2,6,5,1,2])
+    # thinline()
+
+    # twoSumTest([2,7,11,15], 9)
+    # twoSumTest([2,3,4], 6)
+    # twoSumTest([-1,0], -1)
 
 
 # Testing duplicateZeros
@@ -55,3 +109,72 @@ def removeDuplicatesTest(arr: list[int]) -> None:
     print(f"Removing duplicates in {arr}")
     remain = solver.removeDuplicates(arr)
     print(f"Remaining values = {remain}")
+
+# Testing pivotIndex
+def pivotIndexTest(nums: list[int]) -> None:
+    print(f"List = {nums}")
+    answer = solver.pivotIndex(nums)
+    print(f"Pivot index = {answer}")
+
+# Testing dominantIndex
+def dominantIndexTest(nums: list[int]) -> None:
+    print(f"Original list = {nums}")
+    answer = solver.dominantIndex(nums)
+    print(f"Dominant index = {answer}")
+
+# Testing plusOne
+def plusOneTest(nums: list[int]) -> None:
+    print(f"Original List = {nums}")
+    answer = solver.plusOne(nums)
+    print(f"Updated List = {answer}")
+
+# Testing findDiagonalOrder - Doesn't work
+# def findDiagonalOrderTest(matrix: list[list[int]]) -> None:
+#     print(f"Matrix = {matrix}")
+#     answer = solver.findDiagonalOrder(matrix)
+#     print(f"Order = {answer}")
+
+# Testing spiralOrder
+def spiralOrderTest(matrix: list[list[int]]) -> None:
+    print(f"Matrix = {matrix}")
+    answer = solver.spiralOrder(matrix)
+    print(f"Order = {answer}")
+
+# Testing generate
+def generatePascalTest(num: int) -> None:
+    answer = solver.generate(num)
+    print(f"Making a Pascal Triangle from {num} rows = {answer}")
+
+# Testing addBinary
+def binarySumTest(bin1: str, bin2: str) -> None:
+    answer = solver.addBinary(bin1, bin2)
+    print(f"Binary sum of {bin1} + {bin2} = {answer}")
+
+# Testing strStr
+def findStrTest(word: str, compair: str) -> None:
+    print(f"Finding {compair} in {word}")
+    answer = solver.strStr(word, compair)
+    print(f"Index = {answer}")
+
+# Testing longestCommonPrefix
+def longestCommonPrefixTest(words: list[str]) -> None:
+    answer = solver.longestCommonPrefix(words)
+    print(f"Common prefix for {words} is {answer}")
+
+# Testing reverseString
+def reverseStringTest(wordStr: list[str]) -> None:
+    print(f"Original word = {wordStr}")
+    answer = solver.reverseString(wordStr)
+    print(f"Reversed word = {answer}")
+
+# Testing arrayPairSum
+def arrayPairSumTest(nums: list[int]) -> None:
+    print(f"Original list = {nums}")
+    answer = solver.arrayPairSum(nums)
+    print(f"The Max Sum = {answer}")
+
+# Testing twoSum
+# def twoSumTest(nums: list[int], num: int) -> None:
+#     print(f"Original list = {nums}")
+#     answer = solver.twoSum(nums)
+#     print(f"Indecies = {answer}")
